move ssl parts to ssl.go
